requirements.txt
streamlit>=1.49.0
openai>=1.43.0
pydantic>=2.8.2
python-dotenv>=1.0.1
pandas>=2.2.2
requests>=2.32.3
tqdm>=4.66.5
langdetect>=1.0.9
beautifulsoup4>=4.13.0
lxml>=5.3.0
selenium>=4.23.1
undetected-chromedriver>=3.5.5
webdriver-manager>=4.0.2
fake-useragent>=1.5.1
phonenumbers>=8.13.52
dnspython>=2.6.1
email-validator>=2.2.0
httpx>=0.27.2
pytz>=2024.1
2captcha-python
plotly>=5.17.0
asyncio
# Yeni CAPTCHA bypass geliştirmeleri için ek kütüphaneler
urllib3>=2.0.0
certifi>=2023.0.0
charset-normalizer>=3.0.0
idna>=3.0
